+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   respond to `#to_partial_path`, including but not limited to `Link`s.
 * Partials can be ordered
 * Added a batch actions feature
+* Frontend assets can be set up via `yarn add @superadministration/super`
 
 ### Breaking changes